[21.11.16 정민주][Mod] option 관련 함수 분리
